// src/model/replies.ts
export enum BotReplies {
    MASK_EXISTS = "Маска с таким именем уже существует: ",
    MASK_IS_SAVED = "Маска сохранена. Чтобы найти маску, воспользуйтесь командой: /find <UUID_Mask or name>",
    MASK_IS_WRONG = "Неправильный формат маски. Маска может включать следующие символы:" +
        "цифры, буквы a-f в верхнем или нижнем регистре, '?', '-','*'",
    MASK_IS_FOUND = "Искомая маска: ",
    MASKS_ARE_FOUND = "Искомые маски: \n",
    MASK_IS_NOT_FOUND = "Сори, мы не нашли маску. Если хотите создать маску, воспользуйтесь командой: '/create <UUID_Mask><name>'",
    INPUT_MASK = "Пожалуйста введите маску",
    INPUT_NAME = "Пожалуйста введите имя"
}